Dashboard Refactor draft 1
Draft 1 of refactoring dashboard,
Make Tables the dashboard landing page instead
