+++ dbs_frontend_react/src/components/_nav.js
@@ -30,12 +30,6 @@ const _nav = [
   },
   {
     component: CNavItem,
-    name: 'Table',
-    to: '/table',
-    icon: <CIcon icon={cilSpeedometer} customClassName="nav-icon" />,
-  },
-  {
-    component: CNavItem,
     name: 'Order',
     to: '/order',
     icon: <CIcon icon={cilSpeedometer} customClassName="nav-icon" />,
